requirements.txt
# Runtime: Python 3.11 or newer
# Phase 0 도구는 표준 라이브러리만 사용합니다.
# 추후 Markdown 파서나 XML 라이브러리를 도입할 경우 아래에 패키지를 추가하세요.

PyYAML>=6.0
pytest>=8.0
